Extract declarations.
